Added script to find_record.py
